requirements.txt
flask
pytesseract
pillow
python-docx
gunicorn
